bot_docker/trading_robot/pyproject.toml
[tool.poetry]
name = "trading_robot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
start = "^0.2"
robot = "^20071211"
numpy = "^1.20.1"
mysql-connector-python = "^8.0.23"
openapi-client = {git = "https://github.com/andreiandreeev/deribit_cloned.git"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
